refactor: Rename `admin-config` component to `server-settings`
This better separates the component from the `admin-config` service with
the same name. The `server` prefix leaves room for adding
`user-settings` later on. The name `settings` is chosen over `config` to
emphasize that these can be changed at runtime.
