+++ components/server-settings/api-model/build.gradle.kts
