+++ settings.gradle.kts
@@ -28,8 +28,6 @@ include(":api:v1:mapping")
 include(":api:v1:model")
 include(":cli")
 include(":clients:keycloak")
-include(":components:admin-config:api-model")
-include(":components:admin-config:backend")
 include(":components:authorization:api-model")
 include(":components:authorization:backend")
 include(":components:authorization-keycloak:backend")
@@ -47,6 +45,8 @@ include(":components:search:api-model")
 include(":components:search:backend")
 include(":components:secrets:api-model")
 include(":components:secrets:backend")
+include(":components:server-settings:api-model")
+include(":components:server-settings:backend")
 include(":compositions:secrets-routes")
 include(":config")
 include(":config:git")
